requirements.txt
# Core AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# AWS CDK for Infrastructure as Code
aws-cdk-lib>=2.120.0
constructs>=10.3.0

# Security Scanning
semgrep>=1.50.0

# HTTP Client for API integrations
httpx>=0.26.0

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
moto>=4.2.0

# Type Checking
mypy>=1.8.0
types-boto3>=1.0.2

# Code Quality
black>=23.12.0
ruff>=0.1.9

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0

